requirements.txt
Pillow
fonttools
brotli
zopfli